+++ packages/mocks/storybook-addon/preview.js
@@ -1,0 +1,3 @@
+import { withMocks } from '../storybook/addon/decorator.js';
+
+export const decorators = [withMocks];