+++ packages/storybook-framework-web-components/src/preset.ts
@@ -5,3 +5,10 @@ export const core: PresetProperty<'core', StorybookConfig> = {
   builder: '@web/storybook-builder',
   renderer: '@storybook/web-components',
 };
+
+export const wdsFinal: StorybookConfig['wdsFinal'] = wdsConfig => {
+  return {
+    ...wdsConfig,
+    watch: true,
+  };
+};