+++ packages/storybook-utils/index.mjs
@@ -1,1 +1,1 @@
-export * from './dist/index.js';
+export * from './src/index.js';