+++ packages/dev-server-import-maps/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web/dev-server-import-maps
 
+## 0.2.1
+
+### Patch Changes
+
+- fix: update @web/dev-server-core
+
 ## 0.2.0
 
 ### Minor Changes