+++ packages/dev-server-storybook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web/dev-server-storybook
 
+## 2.0.3
+
+### Patch Changes
+
+- fix: update @web/dev-server-core
+
+## 2.0.2
+
+### Patch Changes
+
+- d2dbb7b1: fix: sb windows path
+
 ## 2.0.1
 
 ### Patch Changes