+++ packages/storybook-framework-web-components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web/storybook-framework-web-components
 
+## 0.1.3
+
+### Patch Changes
+
+- c0fefb04: fix storybook-framework-web-components on Windows
+- Updated dependencies [c0fefb04]
+  - @web/storybook-builder@0.1.17
+
+## 0.1.2
+
+### Patch Changes
+
+- 589bad6c: watch code changes and auto-reload by default
+
 ## 0.1.1
 
 ### Patch Changes