+++ packages/test-runner-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web/test-runner-core
 
+## 0.13.3
+
+### Patch Changes
+
+- 39ff6ffb: replace ip dependency due to security bug CVE-2024-29415
+
+## 0.13.2
+
+### Patch Changes
+
+- fix: update @web/dev-server-core
+
 ## 0.13.1
 
 ### Patch Changes