+++ packages/test-runner-module-mocking/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web/test-runner-module-mocking
 
+## 0.1.1
+
+### Patch Changes
+
+- fix: update @web/dev-server-core
+
 ## 0.1.0
 
 ### Minor Changes